Safety commit-m

+++ src/controller/connect_four.py
@@ -74,9 +74,9 @@ class Connect_Four(Game):
 
         # Structure data as a 4x4x2 stack.
         if state.player:
-            return np.array([pos_pieces, neg_pieces])
+            return np.array([pos_pieces, neg_pieces], dtype="float32")
         else:
-            return np.array([neg_pieces, pos_pieces])
+            return np.array([neg_pieces, pos_pieces], dtype="float32")
 
     def map_logits(self, actions, logits):
         action_map = dict()
@@ -93,7 +93,7 @@ class Connect_Four(Game):
         return action_map
 
     def map_actions(self, target_policies):
-        policies = np.zeros((self.size * self.size))
+        policies = np.zeros((self.size * self.size), dtype="float32")
         for a, p in target_policies.items():
             y, x = a.dest
             policies[y * self.size + x] = p
